refactor sign-in \ sign-up view components
