+++ src/app/template/auth/auth-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { AuthComponent } from './component/auth.component';
 import { SignInComponent } from './children/sign-in/component/sign-in.component';
-import { SignUpComponent } from './children/sign-up/sign-up.component';
+import { SignUpComponent } from './children/sign-up/component/sign-up.component';
 import { AuthenticationComponent } from './children/authentication/authentication.component';
 
 const routes: Routes = [
