+++ src/app/template/auth/auth.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 
 import { AuthComponent } from './component/auth.component';
 import { SignInComponent } from './children/sign-in/component/sign-in.component';
-import { SignUpComponent } from './children/sign-up/sign-up.component';
+import { SignUpComponent } from './children/sign-up/component/sign-up.component';
 import { AuthenticationComponent } from './children/authentication/authentication.component';
 import { AuthRoutingModule } from './auth-routing.module';
 import { SharedModule } from '../../shared/shared.module';
