+++ src/app/template/auth/children/sign-in/component/sign-in.component.html
@@ -1,6 +1,5 @@
 <form>
   <div class="sign-in">
-    <div class="sign-in-item head-title-text">sign in</div>
     <div class="sign-in-item">
       <app-input [inputParameters]="signInConfig.inputs.email"></app-input>
     </div>
