+++ src/app/template/auth/children/sign-up/component/sign-up.component.html
@@ -1,0 +1,19 @@
+<form>
+  <div class="sign-in">
+    <div class="sign-in-item">
+      <app-input [inputParameters]="signUpConfig.inputs.firstName"></app-input>
+    </div>
+    <div class="sign-in-item">
+      <app-input [inputParameters]="signUpConfig.inputs.lastName"></app-input>
+    </div>
+    <div class="sign-in-item">
+      <app-input [inputParameters]="signUpConfig.inputs.email"></app-input>
+    </div>
+    <div class="sign-in-item">
+      <app-input [inputParameters]="signUpConfig.inputs.password"></app-input>
+    </div>
+    <div class="sign-in-item">
+      <app-button [buttonParameters]="signUpConfig.buttons.signUp"></app-button>
+    </div>
+  </div>
+</form>
