+++ src/app/template/auth/children/sign-up/component/sign-up.component.scss
