--- src/app/template/auth/children/sign-up/sign-up.component.html
@@ -1,1 +1,0 @@
-<p>sign-up works!</p>
