+++ src/app/template/auth/component/auth.component.scss
@@ -1,0 +1,30 @@
+@import "src/assets/scss/mixins";
+
+:host {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.auth {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  &-head {
+    color: $lynch;
+    margin-bottom: 50px;
+  }
+
+  &-link {
+    color: $lynch;
+  }
+
+  &-link.active {
+    color: $dark;
+    text-decoration: underline;
+  }
+}
